+++ src/components/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import { HomePage, SignUpPage, LoginPage, MeetupsPage, MeetupDetailPage } from '../pages';
+import { Home, SignUp, Login, Meetups, MeetupDetail } from '../pages';
 import { Footer } from './shared';
 import Header from './shared/Header';
 
@@ -9,11 +9,11 @@ const App = () => (
     <Header />
     <Router>
       <Switch>
-        <Route path="/meetups/:id" component={MeetupDetailPage} />
-        <Route path="/meetups" component={MeetupsPage} />
-        <Route path="/login" component={LoginPage} />
-        <Route path="/signup" component={SignUpPage} />
-        <Route path="/" component={HomePage} />
+        <Route path="/meetups/:id" component={MeetupDetail} />
+        <Route path="/meetups" component={Meetups} />
+        <Route path="/login" component={Login} />
+        <Route path="/signup" component={SignUp} />
+        <Route path="/" component={Home} />
       </Switch>
     </Router>
     <Footer />
