+++ src/components/index.ts
@@ -1,11 +1,9 @@
-import HomePageNavBar from './HomePageNavBar';
 import NavBar from './NavBar';
 import UpcomingMeetups from './UpcomingMeetups';
 import InfoCards from './InfoCards';
 import ShowCase from './ShowCase';
 
 export {
-  HomePageNavBar,
   NavBar,
   InfoCards,
   UpcomingMeetups,
