--- src/components/shared/FormButton.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const FormButton = ({ children, classList, disabled }) => (
-  <button
-    type="submit"
-    className={classList}
-    disabled={disabled}
-  >
-    {children}
-  </button>
-);
-
-FormButton.propTypes = {
-  children: PropTypes.node,
-  classList: PropTypes.string
-};
-
-export default FormButton;
