--- src/components/shared/FormGroup.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const FormGroup = ({ classList, children }) => (
-  <div className={classList}>
-    {children}
-  </div>
-);
-
-FormGroup.propTypes = {
-  classList: PropTypes.string,
-  children: PropTypes.node
-};
-
-export default FormGroup;
