--- src/components/shared/__tests__/Form.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import { Form } from '..';
-
-describe('Form Component', () => {
-  test('it renders <Form />', () => {
-    const form = render(<Form />);
-    expect(form).toBeTruthy();
-  });
-});
