+++ src/components/shared/index.ts
@@ -1,14 +1,9 @@
-import Form from './Form';
-import FormGroup from './FormGroup';
-import FormInputField from './FormInputField';
-import FormLabel from './FormLabel';
 import RightNav from './RightNav';
 import AuthRightNav from './AuthRightNav';
 import LeftNav from './LeftNav';
 import Card from './Card';
 import CardTop from './CardTop';
 import CardBottom from './CardBottom';
-import FormButton from './FormButton';
 import MobileNavMenu from './MobileNavMenu';
 import Footer from './Footer';
 import Meetup from './Meetup';
@@ -16,17 +11,12 @@ import Meetups from './Meetups';
 import Dropdown from './Dropdown';
 
 export {
-  Form,
-  FormGroup,
-  FormInputField,
-  FormLabel,
   AuthRightNav,
   RightNav,
   LeftNav,
   Card,
   CardTop,
   CardBottom,
-  FormButton,
   MobileNavMenu,
   Footer,
   Meetup,
